+++ src/main/java/project/onlinebookstore/dto/book/BookSearchParameters.java
@@ -1,4 +1,4 @@
-package project.onlinebookstore.dto;
+package project.onlinebookstore.dto.book;
 
 import java.math.BigDecimal;
