+++ src/main/java/project/onlinebookstore/repository/SpecificationBuilder.java
@@ -1,7 +1,7 @@
 package project.onlinebookstore.repository;
 
 import org.springframework.data.jpa.domain.Specification;
-import project.onlinebookstore.dto.BookSearchParameters;
+import project.onlinebookstore.dto.book.BookSearchParameters;
 
 public interface SpecificationBuilder<T> {
     Specification<T> build(BookSearchParameters searchParameters);
