+++ src/main/java/project/onlinebookstore/service/book/BookService.java
@@ -1,10 +1,10 @@
-package project.onlinebookstore.service;
+package project.onlinebookstore.service.book;
 
 import java.util.List;
 import org.springframework.data.domain.Pageable;
-import project.onlinebookstore.dto.BookDto;
-import project.onlinebookstore.dto.BookSearchParameters;
-import project.onlinebookstore.dto.CreateBookRequestDto;
+import project.onlinebookstore.dto.book.BookDto;
+import project.onlinebookstore.dto.book.BookSearchParameters;
+import project.onlinebookstore.dto.book.CreateBookRequestDto;
 
 public interface BookService {
     BookDto save(CreateBookRequestDto bookRequestDto);
